+++ CHANGES.rst
@@ -14,6 +14,9 @@ Bug Fixes
 Other Changes and Additions
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+- Convolution-based smoothing will now apply a 1D kernel to multi-dimensional fluxes
+  by convolving along the spectral axis only, rather than raising an error.
+
 1.4.1
 -----